args.rs: move PositionRequest for file readability
* Moved some helper functions to become associated
  functins of PositionRequest.
* Increased usage of Self.
* Moved PositionRequest logic and tests into it's one file.
